taskit/taskit-backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
psycopg2-binary==2.9.10
django-cors-headers==4.6.0
python-dotenv==1.0.1
djangorestframework-simplejwt==5.4.0
celery[redis]==5.4.0
redis==5.2.1
pyyaml==6.0.2
